Completed Day06
This appears to be a memory-management issue.  We can get away with it
if we do it naively in a list, treating each fish individually, but 256
is too many days to deal with - I started running out of memory around
day 117.  Instead a quick re-think and I realised that if I grouped the
fish into the "days remaining" buckets, then I could effectively and
efficiently give an upper bound to the processing and make it a _bunch_
faster.